+++ GISPython/AGServerHelperNTLM.py
@@ -3,14 +3,20 @@
      Module for operations with ArcGIS Server services
 """
 
+import os
 import urllib
 import json
 import urllib2
 import ssl
 import urlparse
+import tempfile
+from lxml import etree
 from collections import OrderedDict
 from ntlm import HTTPNtlmAuthHandler
+
 import MyError
+import TimerHelper
+
 
 class AGServerHelperNTLM(object):
     """Class for operations with ArcGIS Server services"""
@@ -21,7 +27,9 @@ class AGServerHelperNTLM(object):
                  ags_admin_url,
                  tool=None,
                  basic=False,
-                 allowunverifiedssl=False):
+                 allowunverifiedssl=False,
+                 token=False
+                 ):
         """Class initialization procedure
 
         Args:
@@ -29,40 +37,76 @@ class AGServerHelperNTLM(object):
             username: ArcGIS Server administrator username
             password: ArcGIS Server administrator password
             ags_admin_url: ArcGIS server rest admin url
-            Tool: GISPython tool (optional)
+            tool: GISPython tool (optional)
             basic: bool indicating that Basic autentification will be used instead of NTLM
+            allowunverifiedssl: bool indicating that unsecure SSL connections are allowed
+            token: bool indicating that token authentication is to be used
         """
         self.username = username
         self.password = password
         self.ags_admin_url = ags_admin_url
-        self.serverurl = ags_admin_url
+        self.server_url = ags_admin_url
+
         if self.ags_admin_url.endswith("/"):
             self.ags_admin_url = self.ags_admin_url[:-1]
+
         self.Tool = tool
 
         self.ags_admin_url = urlparse.urljoin(self.ags_admin_url, '/arcgis/admin')
 
         if allowunverifiedssl:
-            try:
-                _create_unverified_https_context = ssl._create_unverified_context
-            except AttributeError:
-                # Legacy Python that doesn't verify HTTPS certificates by default
-                pass
-            else:
-                # Handle target environment that doesn't support HTTPS verification
-                ssl._create_default_https_context = _create_unverified_https_context
+            self._set_allow_unsafe_ssl()
 
+        self.authentication_mode = 'ntlm'
+        if basic:
+            self.authentication_mode = 'basic'
+        if token:
+            self.authentication_mode = 'token'
+
+        if self.authentication_mode in ('ntlm', 'basic'):
+            self._set_autentification_handler()
+        else:
+            self.token = self._generate_tocken()
+            if self.token == "":
+                raise MyError.MyError("Could not generate a token with the username and password provided.")
+
+    def _set_autentification_handler(self):
         passman = urllib2.HTTPPasswordMgrWithDefaultRealm()
         passman.add_password(None, self.ags_admin_url, self.username, self.password)
-
-        if not basic:
+        if self.authentication_mode == 'ntlm':
             auth_handler = HTTPNtlmAuthHandler.HTTPNtlmAuthHandler(passman)
         else:
             auth_handler = urllib2.HTTPBasicAuthHandler(passman)
-
         opener = urllib2.build_opener(auth_handler)
         urllib2.install_opener(opener)
 
+    @staticmethod
+    def _set_allow_unsafe_ssl():
+        try:
+            _create_unverified_https_context = ssl._create_unverified_context
+        except AttributeError:
+            # Legacy Python that doesn't verify HTTPS certificates by default
+            pass
+        else:
+            # Handle target environment that doesn't support HTTPS verification
+            ssl._create_default_https_context = _create_unverified_https_context
+
+    def _generate_tocken(self, expiration=60 * 24):
+        query_dict = {
+            'username': self.username,
+            'password': self.password,
+            'expiration': expiration,
+            'client': 'requestip'}
+
+        query_string = urllib.urlencode(query_dict)
+        url = "{}/generateToken".format(self.ags_admin_url)
+
+        token = json.loads(urllib.urlopen(url + "?f=json", query_string).read())
+
+        if "token" not in token:
+            raise MyError.MyError(token['messages'])
+        else:
+            return token['token']
 
     def __request_from_server(self, adress, params, content_type='application/json', method="POST"):
         """Function for ntlm request creation
@@ -77,8 +121,21 @@ class AGServerHelperNTLM(object):
         Returns:
             Response string
         """
+        if self.authentication_mode == 'token':
+            if method == "POST":
+                if params is None:
+                    params = {'token': self.token}
+                else:
+                    params['token'] = self.token
+
+            if '?' in adress:
+                adress = u'{}&token={}'.format(adress, self.token)
+            else:
+                adress = u'{}?token={}'.format(adress, self.token)
+
         data = urllib.urlencode(params)
         url_address = urlparse.urljoin(self.ags_admin_url + "/", adress)
+        print(url_address)
         req = urllib2.Request(url=url_address, data=data)
         req.add_header('Content-Type', content_type)
         req.get_method = lambda: method
@@ -89,32 +146,16 @@ class AGServerHelperNTLM(object):
                 url_address, data, method, response.msg))
         return response_string
 
-    def __assert_json_success(self, data):
-        """Function for aserting json request state
-
-        Args:
-            self: The reserved object 'self'
-            data: Request response string
-
-        Returns:
-            boolean False if request has errors
-        """
+    @staticmethod
+    def __assert_json_success(data):
         obj = json.loads(data)
         if 'status' in obj and obj['status'] == "error":
             raise MyError.MyError("Error: JSON object returns an error. " + str(obj))
         else:
             return True
 
-    def __process_folder_string(self, folder):
-        """Function for processing folder name string
-
-        Args:
-            self: The reserved object 'self'
-            folder: folder string
-
-        Returns:
-            corrected folder string
-        """
+    @staticmethod
+    def __process_folder_string(folder):
         if folder is None:
             folder = "ROOT"
         if folder.upper() == "ROOT":
@@ -150,16 +191,15 @@ class AGServerHelperNTLM(object):
         if not self.__assert_json_success(data):
             raise MyError.MyError("Error when reading folder information. " + str(data))
         else:
-            if self.Tool != None:
-                self.Tool.AddMessage("Service {}{} {}  done successfully ...".format(
-                    folder, service, action))
+            self._add_message("Service {}{} {}  done successfully ...".format(
+                folder, service, action))
 
     def getServiceList(self, folder, return_running_state=True):
         """Retrieve ArcGIS server services
 
         Args:
-            self: The reserved object 'self'
             folder: Folder of the service (ROOT for root services)
+            return_running_state: If False no service running state will be checked (works faster)
         """
         services = []
         folder = self.__process_folder_string(folder)
@@ -175,10 +215,10 @@ class AGServerHelperNTLM(object):
         for single in service_list["services"]:
             services.append(folder + single['serviceName'] + '.' + single['type'])
 
-        folder_list = service_list["folders"] if u'folders' in service_list != False else []
-        if u'Utilities' in service_list != False:
+        folder_list = service_list["folders"] if u'folders' in service_list else []
+        if u'Utilities' in service_list:
             folder_list.remove("Utilities")
-        if u'System' in service_list != False:
+        if u'System' in service_list:
             folder_list.remove("System")
 
         if folder_list:
@@ -191,21 +231,17 @@ class AGServerHelperNTLM(object):
                     services.append(subfolder + "/" + single['serviceName'] + '.' + single['type'])
 
         if not services:
-            if self.Tool != None:
-                self.Tool.AddMessage("No services found")
+            self._add_message("No services found")
         else:
-            if self.Tool != None:
-                self.Tool.AddMessage("Services on " + self.serverurl +":")
+            self._add_message("Services on " + self.server_url + ":")
             for service in services:
                 if return_running_state:
                     status_url = "services/{}/status?f=json".format(service)
                     data = self.__request_from_server(status_url, params)
                     status = json.loads(data)
-                    if self.Tool != None:
-                        self.Tool.AddMessage("  " + status["realTimeState"] + " > " + service)
+                    self._add_message("  " + status["realTimeState"] + " > " + service)
                 else:
-                    if self.Tool != None:
-                        self.Tool.AddMessage(" > " + service)
+                    self._add_message(" > " + service)
 
         return services
 
@@ -214,7 +250,7 @@ class AGServerHelperNTLM(object):
 
         Args:
             self: The reserved object 'self'
-            serverService: Service which parameter configuration shall be retrieved
+            server_service: Service which parameter configuration shall be retrieved
 
         Returns:
             json data object
@@ -227,13 +263,12 @@ class AGServerHelperNTLM(object):
             raise MyError.MyError(
                 u'...Couldn\'t retrieve service parameter configuration: {}\n'.format(data))
         else:
-            if self.Tool != None:
-                self.Tool.AddMessage(u'...Service parameter configuration successfully retrieved\n')
+            self._add_message(u'...Service parameter configuration successfully retrieved\n')
 
         data_object = json.loads(data)
         return data_object
 
-    def publishServerJson(self, service, data_object):
+    def publishServerJson(self, service, data_object, gp=None):
         """Publish service parameters to server
 
         Args:
@@ -251,12 +286,10 @@ class AGServerHelperNTLM(object):
                                                method='POST')
 
         if not self.__assert_json_success(edit_data):
-            if self.Tool != None:
-                self.Tool.AddMessage(
-                    u'...Service configuration renewal error: {}\n'.format(edit_data))
+            self._add_message(
+                u'...Service configuration renewal error: {}\n'.format(edit_data), gp)
         else:
-            if self.Tool != None:
-                self.Tool.AddMessage(u'...Service configuration succesfully renewed\n')
+            self._add_message(u'...Service configuration succesfully renewed\n')
 
         return
 
@@ -269,11 +302,11 @@ class AGServerHelperNTLM(object):
             service: Name of the service from which to get corresponding name from the server services list
             serviceDir: Name of the service directory which is shown in the configuration of services to be published on the server
         """
-        server_service = ''
         if serviceDir is None:
             config_service = service
         else:
             config_service = serviceDir + "/" + service
+
         for server_service in services:
             if server_service.split('.')[0].upper() == config_service.upper():
                 return server_service
@@ -303,8 +336,8 @@ class AGServerHelperNTLM(object):
         return_dict = OrderedDict()
         for permision in statusdata_object["permissions"]:
             principal = permision["principal"]
-            if permision.has_key("permission"):
-                if permision["permission"].has_key("isAllowed"):
+            if "permission" in permision:
+                if "isAllowed" in permision["permission"]:
                     return_dict[principal] = permision["permission"]["isAllowed"]
                 else:
                     return_dict[principal] = True
@@ -320,7 +353,7 @@ class AGServerHelperNTLM(object):
             principal: The name of the role for whom the permission is being assigned.
             is_allowed: Tells if access to a resource is allowed or denied.
         """
-        urlparams = urllib.urlencode({'principal':principal, 'isAllowed':is_allowed, 'f':'json'})
+        urlparams = urllib.urlencode({'principal': principal, 'isAllowed': is_allowed, 'f': 'json'})
         folder = self.__process_folder_string(folder)
         status_url = "services/" + folder + service + "/permissions/add?{}".format(urlparams)
         params = {}
@@ -387,6 +420,7 @@ class AGServerHelperNTLM(object):
 
         Args:
             self: The reserved object 'self'
+            pageSize: The maximums records rturned
 
         Returns: list of strings
         """
@@ -400,12 +434,11 @@ class AGServerHelperNTLM(object):
 
         roles = json.loads(roles_data)
         for role in roles['roles']:
-            if role.has_key("description"):
+            if "description" in role:
                 rezult.append({"rolename": role["rolename"],
                                "description": role["description"]})
             else:
                 rezult.append({"rolename": role["rolename"]})
-
 
         return rezult
 
@@ -416,7 +449,7 @@ class AGServerHelperNTLM(object):
             self: The reserved object 'self'
             rolename: The name of the role. The name must be unique in the role store.
         """
-        urlparams = urllib.urlencode({'rolename':rolename, 'description':description, 'f':'json'})
+        urlparams = urllib.urlencode({'rolename': rolename, 'description': description, 'f': 'json'})
         manifest_url = "security/roles/add?{}".format(urlparams)
         params = {}
         roles_data = self.__request_from_server(manifest_url, params, method='POST')
@@ -431,7 +464,7 @@ class AGServerHelperNTLM(object):
             self: The reserved object 'self'
             rolename: The name of the role.
         """
-        urlparams = urllib.urlencode({'rolename':rolename, 'f':'json'})
+        urlparams = urllib.urlencode({'rolename': rolename, 'f': 'json'})
         manifest_url = "security/roles/remove?{}".format(urlparams)
         params = {}
         roles_data = self.__request_from_server(manifest_url, params, method='POST')
@@ -444,10 +477,11 @@ class AGServerHelperNTLM(object):
 
         Args:
             self: The reserved object 'self'
+            maxCount: maximum returned record count
 
         Returns: list of strings
         """
-        urlparams = urllib.urlencode({'rolename':rolename, 'maxCount':maxCount, 'f':'json'})
+        urlparams = urllib.urlencode({'rolename': rolename, 'maxCount': maxCount, 'f': 'json'})
         manifest_url = "security/roles/getUsersWithinRole?{}".format(urlparams)
         params = {}
         users_data = self.__request_from_server(manifest_url, params, method='POST')
@@ -470,7 +504,7 @@ class AGServerHelperNTLM(object):
             rolename: The name of the role.
             users: A comma-separated list of user names. Each user name must exist in the user store.
         """
-        urlparams = urllib.urlencode({'rolename':rolename, 'users':users, 'f':'json'})
+        urlparams = urllib.urlencode({'rolename': rolename, 'users': users, 'f': 'json'})
         manifest_url = "security/roles/addUsersToRole?{}".format(urlparams)
         params = {}
         roles_data = self.__request_from_server(manifest_url, params, method='POST')
@@ -487,7 +521,7 @@ class AGServerHelperNTLM(object):
             rolename: The name of the role.
             users: A comma-separated list of user names. Each user name must exist in the user store.
         """
-        urlparams = urllib.urlencode({'rolename':rolename, 'users':users, 'f':'json'})
+        urlparams = urllib.urlencode({'rolename': rolename, 'users': users, 'f': 'json'})
         manifest_url = "security/roles/removeUsersFromRole?{}".format(urlparams)
         params = {}
         roles_data = self.__request_from_server(manifest_url, params, method='POST')
@@ -518,7 +552,7 @@ class AGServerHelperNTLM(object):
         statusdata_object = json.loads(status_data)
         for database in statusdata_object['databases']:
             dataset_names = [d['onServerName'] for d in database['datasets']]
-            item = {"database": database['onServerName'], "datasets":dataset_names}
+            item = {"database": database['onServerName'], "datasets": dataset_names}
             rezult.append(item)
 
         return rezult
@@ -582,3 +616,207 @@ class AGServerHelperNTLM(object):
                 rezult.append(statusdata_object)
 
         return rezult
+
+    def publish_mxd(self, mxd, service, service_folder, arcgis_server_connection, gp, create_new=False,
+                    aditional_params={}):
+        """Publishes MXD to server
+
+        Args:
+            self: The reserved object 'self'
+            mxd: MXD document name
+            service: AGS service name
+            service_folder: AGS service folder
+            create_new: create new AGS service?
+            arcgis_server_connection: path to ArcGIS Server connection file *.ags
+            gp: ArcPy object
+            create_new: bool True if new service can be created if service does not exist
+            aditional_params: (example)
+                [
+                    {
+                        'TypeName': 'MapServer',
+                        'ConfigurationProperties': {
+                            'enableDynamicLayers': 'false',
+                            'maxDomainCodeCount': '250000'
+
+                        },
+                        "Props": {
+                            'MaxInstances': '10'
+                        }
+                    },
+                    {
+                        'TypeName': 'KmlServer',
+                        'Enabled': 'false'
+                    },
+                    {
+                        'TypeName': 'FeatureServer',
+                        'Enabled': 'true',
+                        "Props": {
+                            'maxRecordCount': '2000'
+                        }
+                    }
+                ]
+        """
+        service_exists = True
+
+        server_admin_temp = tempfile.gettempdir()
+        mxd_name = os.path.splitext(os.path.basename(mxd))[0]
+        sdd_draft = os.path.join(server_admin_temp, mxd_name) + '.sddraft'
+        sd = os.path.join(server_admin_temp, mxd_name) + '.sd'
+
+        timer = TimerHelper.TimerHelper()
+
+        self._add_message(
+            u'\n...Processing MXD file [{}] and preparing service for publish\n'.format(mxd), gp)
+
+        services = self.getServiceList(service_folder, False)
+        server_service = self.getServiceFromServer(services, service, service_folder)
+
+        if server_service == '':
+            service_exists = False
+
+        if not service_exists and not create_new:
+            self._add_warning(
+                u'\n===================================\n...Service [' + service + u'] does not exist on server!\n===================================\n')
+            return False
+
+        json_data = self.GetServerJson(server_service)
+
+        self._add_message(u'...Creating SDDraft: {} - {}\n'.format(sdd_draft, timer.GetTimeReset()), gp)
+
+        mxd = gp.mapping.MapDocument(mxd)
+        gp.mapping.CreateMapSDDraft(
+            mxd, sdd_draft, service, 'ARCGIS_SERVER',
+            os.path.join(server_admin_temp, arcgis_server_connection), False, service_folder)
+
+        gp.env.overwriteOutput = True
+        self._correct_sdd_draft_configuration(sdd_draft, service_exists, aditional_params)
+
+        self._add_message(u'...Start SDDraft validation {}\n'.format(timer.GetTimeReset()), gp)
+        analysis = gp.mapping.AnalyzeForSD(sdd_draft)
+
+        error_code = self._print_sdd_analysis_results(analysis, sdd_draft, gp)
+
+        if analysis['errors'] == {} and error_code != 24011 and error_code != 24012:
+            self._add_message(
+                u'...SDDraft validation no critical problems where found - {}\n'.format(timer.GetTimeReset()), gp)
+            self._add_message(u'...Start creating stage service \n', gp)
+
+            gp.StageService_server(sdd_draft, sd)
+
+            self._add_message(u'...Start publishing service {}\n'.format(timer.GetTimeReset()), gp)
+            gp.UploadServiceDefinition_server(sd, os.path.join(server_admin_temp, arcgis_server_connection))
+
+            if not service_exists:
+                self._add_message(
+                    u'...Server did not contain service, and service has been created: {}'.format(service), gp)
+                return True
+            else:
+                self._renew_service_configuration(json_data, server_service, timer, gp)
+                return True
+        else:
+            self._add_message(
+                u'...SDDraft file validation critical errors found {}'.format(timer.GetTimeReset()), gp)
+            return False
+
+    @staticmethod
+    def _correct_sdd_draft_configuration(sdd_draft, service_exists, aditional_params):
+        xml = etree.parse(sdd_draft)
+        # xml.write(sdd_draft, pretty_print=True, method="xml", xml_declaration=True, encoding="utf-8")
+
+        if service_exists:
+            xml.xpath('/SVCManifest/Type')[0].text = 'esriServiceDefinitionType_Replacement'
+        xml.xpath('/SVCManifest/State')[0].text = 'esriSDState_Published'
+
+        if aditional_params is not None and aditional_params:
+            for extension_config in aditional_params:
+                if extension_config['TypeName'] == 'MapServer':
+                    extension_tag = xml.xpath(
+                        '/SVCManifest/Configurations/SVCConfiguration/TypeName[text()="{}"]/..'.format(
+                            extension_config['TypeName']))[0]
+                    if 'ConfigurationProperties' in extension_config:
+                        for config_property, property_value in extension_config['ConfigurationProperties'].items():
+                            extension_tag.xpath(
+                                './Definition/ConfigurationProperties/PropertyArray/PropertySetProperty/Key[text()="{}"]/../Value'.format(
+                                    config_property))[0].text = property_value
+                    if 'Props' in extension_config:
+                        for config_property, property_value in extension_config['Props'].items():
+                            extension_tag.xpath(
+                                './Definition/Props/PropertyArray/PropertySetProperty/Key[text()="{}"]/../Value'.format(
+                                    config_property))[0].text = property_value
+                else:
+                    extension_tag = xml.xpath(
+                        '/SVCManifest/Configurations/SVCConfiguration/Definition/Extensions/SVCExtension/TypeName[text()="{}"]/..'.format(
+                            extension_config['TypeName']))[0]
+                    if 'Enabled' in extension_config:
+                        extension_tag.xpath('./Enabled')[0].text = extension_config['Enabled']
+                    if 'Props' in extension_config:
+                        for config_property, property_value in extension_config['Props'].items():
+                            extension_tag.xpath(
+                                './Props/PropertyArray/PropertySetProperty/Key[text()="{}"]/../Value'.format(
+                                    config_property))[0].text = property_value
+
+        xml.write(sdd_draft, pretty_print=True, method="xml", xml_declaration=True, encoding="utf-8")
+
+    def _print_sdd_analysis_results(self, analysis, sdd_draft, gp):
+        error_code = 0
+        iterable_keys = [key for key in ('messages', 'warnings', 'errors') if analysis[key] != {}]
+        for key in iterable_keys:
+            self._add_message("---" + key.upper() + "---", gp)
+            collection = analysis[key]
+            for ((message, code), layerlist) in collection.iteritems():
+                laystr = ', '.join([layer.name for layer in layerlist])
+                text = u"    [{0}] - {1}".format(code, message)
+                if unicode(laystr) != '':
+                    text = text + u"Applies to: ({0})".format(laystr)
+                if key == 'messages':
+                    self._add_message(text, gp)
+                if key == 'warnings':
+                    self._add_message(text, gp)
+                    if code == 24011 or code == 24012:
+                        error_code = code
+                        self._add_warning(
+                            u'...\n===================================================\nData source not reggistered on server. To prevent data copy on server - publish will stop: {}\n==================================================='.format(
+                                sdd_draft), gp)
+                if key == 'errors':
+                    self._add_error(text)
+        return error_code
+
+    def _renew_service_configuration(self, json_data, server_service, timer, gp):
+        self._add_message(u'...read new config from server - {}\n'.format(timer.GetTimeReset()))
+        new_json_data = self.GetServerJson(server_service)
+        json_data["datasets"] = new_json_data["datasets"]
+        self.publishServerJson(server_service, json_data, gp)
+        self._add_message(
+            u'...Service configuration on server updated {}'.format(timer.GetTimeReset()), gp)
+
+    @staticmethod
+    def _set_tag_data_values(doc, tag_name, parent_tag_name, data):
+        type_tags = doc.getElementsByTagName(tag_name)
+        for tag in type_tags:
+            if tag.parentNode.tagName == parent_tag_name:
+                if tag.hasChildNodes():
+                    tag.firstChild.data = data
+
+    def _add_message(self, message, gp=None):
+        if self.Tool is not None:
+            self.Tool.AddMessage(message)
+        elif gp is not None:
+            gp.AddMessage(message)
+        else:
+            print message
+
+    def _add_warning(self, message, gp=None):
+        if self.Tool is not None:
+            self.Tool.AddWarning(message)
+        elif gp is not None:
+            gp.AddWarning(message)
+        else:
+            print message
+
+    def _add_error(self, message, gp=None):
+        if self.Tool is not None:
+            self.Tool.AddError(message)
+        elif gp is not None:
+            gp.AddError(message)
+        else:
+            print message