+++ GISPython/TimerHelper.py
@@ -75,6 +75,7 @@ class TimedSubprocess:
         """
         self.Tool.AddMessage(u'\n\n{0}------------------------\n{0}>>>> Begin {1} - {2}\n{0}------------------------\n'.format(self.prefix, self.txt, self.Tool.MyNow()))
         return self
+        
     def __exit__(self, type, value, traceback):
         """With statement closing procedure
         Args: